organized test cases as real time cases that we would like to monitor
