+++ __tests__/loggedin/recieved/chatModalClosedSocket.test.js
@@ -1,13 +1,8 @@
-import { fetchAppData } from '../src/utility'; // Import the function to mock
+import { fetchAppData } from '../../../src/utility'; // Import the function to mock
+
+
 // import { initialize, global_bucket, handleMsgUpdatedEvent } from '../src/index'; // Adjust the import path
-import * as index from '../src/index';  // Import the entire module
-
-// Mock the module
-jest.mock('../src/utility', () => ({
-  fetchAppData: jest.fn(), // Mock the fetchAppData function
-}));
-
-
+import * as index from '../../../src/index';  // Import the entire module
 
 
 import io from "socket.io-client";
@@ -16,18 +11,27 @@ import io from "socket.io-client";
 
 // Now move jest.spyOn() outside of jest.mock()
 
+
+// Mock the module
+jest.mock('../../../src/utility', () => ({
+  fetchAppData: jest.fn(), // Mock the fetchAppData function
+}));
+
+
+
+
+
+
 // Mock the socket object returned by io()
 jest.mock('socket.io-client', () => {
   const actualSocket = jest.requireActual('socket.io-client')();
   const mSocket = {
     ...actualSocket,
     emit: jest.fn((event, ...args) => {
-      console.log(`emit called with emit event: ${event}, args:`, args);
+      console.log(`emit called with event: ${event}, args:`, args);
       // actualSocket.emit(event, ...args); // Optionally call the real implementation if needed
     }),
     on: jest.fn((event, callback) => {
-      console.log(`.on called with on event: ${event}, callback:`, callback);
-
       // Store the callback if it's not ON_MESSAGE_ARRIVAL_BOT
       mSocket[event] = callback;
       
@@ -36,13 +40,12 @@ jest.mock('socket.io-client', () => {
       }
       
       if (event === 'ON_MESSAGE_STATUS_CHANGED') {
-        console.log("aerweqrwer")
 
         actualSocket.on(event, callback);
       }
       else {
         // You can also add logging here if needed
-        console.log(`on called with elseCase: ${event}`);
+        console.log(`on called with event: ${event}`);
       }
     }),
   };
@@ -52,7 +55,7 @@ jest.mock('socket.io-client', () => {
 
 
 
-describe('Socket Tests2 when window is open', () => {
+describe('Socket Tests', () => {
 
   beforeEach(() => {
     jest.clearAllMocks(); // Clears the call history and reset mock implementations
@@ -65,8 +68,7 @@ describe('Socket Tests2 when window is open', () => {
 
   index.setUp("app1_acm_true_tenant5","dGVuYW50NV9fU0VQUkFUT1JfX2FwcDFfYWNtX3RydWVfdGVuYW50NQ==")
 
-  
-  test('If chat_modal is OPEN :: if message is recieved it should go to CHAT_BODY',async() => {
+  test('If chat_modal is closed :: if message is recieved it should store in the global_bucket', () => {
     // Mock loggedInUser to simulate a logged-in state
     const loggedInUser = {
       tenant: 'tenant1',
@@ -86,70 +88,48 @@ describe('Socket Tests2 when window is open', () => {
     localStorage.setItem('tezkit_token', 'eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpc3MiOiJjb20uemFsYW5kby5jb25uZXhpb24iLCJpYXQiOjE3MjE4Mjk5MTAsImV4cCI6MTcyNzgyOTkxMCwidXNlcl9pZCI6IjMiLCJ1c2VyX3R5cGUiOiJvdGhlciIsImVtYWlsIjoidTJAZ21haWwuY29tIiwidGVuYW50X2FjY291bnRfbmFtZSI6InRlbmFudDEiLCJyb2xlX3BvbGljeSI6Ilt7XCJyb2xlXCI6IDY1NTM2LjAsIFwiYXBwX25hbWVcIjogXCJteW5ld2FwcDJcIn1dIn0.QmhuPBHKFO37BnVJDDtTYd013NoObA_ZI-ppio3NT8o');
     localStorage.setItem('tezkit_app_data', JSON.stringify({"app_id":"NDNxXp0K8i","auth_key":"dGVuYW50NV9fU0VQUkFUT1JfX2FwcDFfYWNtX3RydWVfdGVuYW50NQ==","settings":{"authCloudManaged":true,"requestCloudManaged":null,"app_type":"p2a"},"role":"Owner","tenant_id":"tenant5","timestamp":1728500149,"app_name":"app1_acm_true_tenant5","region":"IN"}));
 
-
     // Call initialize function with loggedInUser
-    await index.initialize(loggedInUser);
-
-
-
-     
- 
- 
+    index.initialize(loggedInUser);
 
     // Get the mock socket instance
     const mockSocket = io();
 
-
-    
     // Verify that socket.emit was called with the correct arguments
     const expectedPayload = { room: 'global_for__2' }; // Adjust as needed
     expect(mockSocket.emit).toHaveBeenCalledTimes(1);
     expect(mockSocket.emit).toHaveBeenCalledWith('join_room', expectedPayload);
-
-
-
-
 
     // Example of manually triggering the 'ON_MESSAGE_ARRIVAL_BOT' event
     if (mockSocket['ON_MESSAGE_ARRIVAL_BOT']) {
       const payload = {
         "message": {
           "msg_id": "msg_id__1",
+          "frm_user":{"id":1},
           "message": "sdfsadf",
           "status": "SENT",
           "timestamp": "23:30:21",
           "to_user": {
             "user": "user2user"
-          },
-          "frm_id": "1"
+          }
         }
       };
-      // expect(index.global_bucket.unread_msgs).toHaveLength(0);
-
-
-      const chatOpener = document.getElementById('chat_modal_opener');
-      const fa_icon = chatOpener.firstChild;
-      fa_icon.click(); // Trigger click event
-
-
-
-      const chatBody1 = document.getElementById('chatBody');
-      console.log("waewrewsdafsdf", chatBody1.children.length)
-
-      expect(chatBody1.children.length).toEqual(0)
       mockSocket['ON_MESSAGE_ARRIVAL_BOT'](JSON.stringify(payload));
-      expect(mockSocket.emit).toHaveBeenCalledTimes(3);
-
-      //all five major listeners are on
-      expect(mockSocket.on).toHaveBeenCalledTimes(5);
-      expect(chatBody1.children.length).toEqual(1)
+      expect(mockSocket.emit).toHaveBeenCalledTimes(2);
 
 
-      
-      
+      expect(index.global_bucket.unread_msgs).toHaveLength(1);
+
+
+      //**************  WE CAN EXTEND IT FUTHER TO SEE IF THE MESSAGES REACHED TO THE DESIRED WINDOW?????****************** */
+
+
+
+
     }
 
-    
-      
+    expect(mockSocket.emit).toHaveBeenCalledTimes(2);
+
+
   });
 });
+
